+++ app/src/main/java/com/iiap/deeplarva/domain/response/CameraConfigurationResponse.kt
@@ -2,6 +2,6 @@ package com.iiap.deeplarva.domain.response
 
 data class CameraConfigurationResponse (
     val iso: Int,
-    val exposure: Long,
+    val exposure: Float,
     val shutterSpeed: Long,
 )
